[outil-csv]refacto: déplacement de makeSendingCSVs et dépendances dans leur propre fichier
