+++ outil-csv.html
@@ -47,7 +47,7 @@
         <script src="https://d3js.org/d3-fetch.v1.min.js"></script>
         <script src="https://d3js.org/d3-dsv.v1.min.js"></script>
         
-        <script defer src="assets/js/main.js"></script>
+        <script type="module" src="assets/js/main.js"></script>
 
     </head>
     <body>
